Add support for read groups in alignment files.
You can now get metrics for every read group in your alignment file. The
semantics of some arguments have changed a bit so that they either
override metadata inferred from read group tags, or provide defaults
when read groups are not used.

For example, --peak-file can now be set to "auto" to look for peak files
named after read groups, or be set to a single filename to use for all
alignments.

Please review the output of 'ataqc --help' for details.

The handling of organisms and reference genomes has changed. Now our
default sets of autosomal references are based on organism, not a
particular reference, and contain both prefixed ("chr1") and
numeric ("1") chromosome reference names.