Rename project. Major UI update to the web viewer.
Renamed the project to ataqv, to avoid confusion with the ataqc tool
from the Kundaje lab.

Reorganized tabs.

Tables are now handled with DataTables, improving performance and
usability with a large number of experiments.

Plots are now made with D3.

Added plot showing how the fragment length distribution of each
experiment diverges from a reference. The reference distribution can be
supplied in a JSON file, calculated as the average of all the
experiments, or you can just use SRR891268, a sample from the original
ATAC-seq paper.

The fragment length distance of each experiment is also listed in the
quality indicators table.

Added the ability to smooth the fragment length distribution plot, and
switch its y axis scale between linear and exponential. The exponential
scale can reveal nucleosomal periodicity in the tails of the plots.

Plot highligting is coordinated; mouse over a plot or legend and the
same experiment or sample will be highlighted in all plots.

Added the ability to specify the range of the fragment length
distributions when generating a web viewer instance.

Read groups in the alignment file can now be ignored.

Thanks to Peter Orchard for implementing the initial fragment length
distance calculation, and the hours of discussion, coding and testing
that led to the current version of the plot.