Fixes for Mac, newer HTSlib
Changes to get it to build and test on Macs, or on Linux with HTSlib
1.4 or newer.

Macs have /usr/bin/true, not /bin/true.

HTSlib has added JSON support, which introduced a namespace clash, and
data URLs passed to sam_open now require a comma before the content;
the old HTSlib format has been kicked to the curb.